removing duplication
