+++ scan_and_capture/camera.py
@@ -73,13 +73,60 @@ def Camera(Picamera):
         # Close plot, delete tmp file and turn off illumination
         plt.close()
 
+    @classmethod
+    def rfi_preview(cls):
 
+        """
+        Displays the preview overlay.
 
-    def preview():
-        pass
+        This method starts a camera preview as an overlay on the Pi's primary
+        display (HDMI or composite). A :class:`PiRenderer` instance (more
+        specifically, a :class:`PiPreviewRenderer`) is constructed with the
+        keyword arguments captured in *options*, and is returned from the
+        method (this instance is also accessible from the :attr:`preview`
+        attribute for as long as the renderer remains active).  By default, the
+        renderer will be opaque and fullscreen.
 
-    def stop_and_flush():
-        pass
+        This means the default preview overrides whatever is currently visible
+        on the display. More specifically, the preview does not rely on a
+        graphical environment like X-Windows (it can run quite happily from a
+        TTY console); it is simply an overlay on the Pi's video output. To stop
+        the preview and reveal the display again, call :meth:`stop_preview`.
+        The preview can be started and stopped multiple times during the
+        lifetime of the :class:`PiCamera` object.
+
+        All other camera properties can be modified "live" while the preview is
+        running (e.g. :attr:`brightness`).
+
+        .. note::
+
+            Because the default preview typically obscures the screen, ensure
+            you have a means of stopping a preview before starting one. If the
+            preview obscures your interactive console you won't be able to
+            Alt+Tab back to it as the preview isn't in a window. If you are in
+            an interactive Python session, simply pressing Ctrl+D usually
+            suffices to terminate the environment, including the camera and its
+            associated preview.
+        """
+
+        camera = PiCamera()
+        camera.rotation = 180
+
+        frame = 1
+        while True:
+            try:
+                camera.start_preview(alpha=150)
+
+            except KeyboardInterrupt:
+                camera.stop_preview()
+                camera.close()
+                break
+
+    def close_and_flush():
+        # https://forums.raspberrypi.com/viewtopic.php?t=152239
+        camera = PiCamera()
+        camera.stop_preview()  # review later
+
 
 
 if __name__=='__main__':
@@ -95,7 +142,7 @@ if __name__=='__main__':
         camera.take_photo(args.save, args.fpath)
 
     if sys.argv[1] == 'preview':
-        camera.preview()
+        camera.rfi_preview()
 
     if sys.argv[1] == 'off':
         camera.close_and_flush()
